fix ShippingController

+++ src/main/java/project/phoneshop/service/Impl/OrderServiceImpl.java
@@ -68,7 +68,7 @@ public class OrderServiceImpl implements OrderService {
     }
 
     @Override
-    public long countOrderPrice() {
+    public double countOrderPrice() {
         return orderRepository.countPrice();
     }
